+++ hit/service.go
@@ -7,7 +7,7 @@ import (
 
 type Storage interface {
 	InsertHit(ctx context.Context, item Hit) error
-	GetHits(ctx context.Context, endpointID int, from time.Time, to time.Time) ([]Hit, error)
+	GetHits(ctx context.Context, endpointID int64, from time.Time, to time.Time) ([]Hit, error)
 }
 
 type HitService struct {
@@ -24,6 +24,6 @@ func (s *HitService) Insert(ctx context.Context, item Hit) error {
 	return s.storage.InsertHit(ctx, item)
 }
 
-func (s *HitService) Get(ctx context.Context, endpointID int, from time.Time, to time.Time) ([]Hit, error) {
+func (s *HitService) Get(ctx context.Context, endpointID int64, from time.Time, to time.Time) ([]Hit, error) {
 	return s.storage.GetHits(ctx, endpointID, from, to)
 }
